requirements.txt
tensorflow
numpy
opencv-python
mysql-connector-python